Added configurable indexer callback
HALucinator uses this to avoid requiring the Python, gtirb, and pprinter
dependencies by running a similar script in a docker container that has
those deps installed.

Additionally, added the --force-remote option for gtirb_lsp_server. This
allows running the LSP server in STDIO mode even in situations where the
server may not have access to the client's filesystem. This was needed
to enable running STDIO mode through Docker for HALucinator, but it can
also be useful for running the server through SSH or as a different user
on the same system.